Remove -Wshorten-64-to-32 for parser and translator

+++ src/parser/State.hpp
@@ -42,7 +42,7 @@ public:
 
     // openelements methods
     int size() const {
-        return openelements.size();
+      return (int)openelements.size();
     }
 
     void push(int id) {
